+++ R/statworx_palette.R
@@ -3,9 +3,12 @@
 #' @description
 #'  A palette based on STATWORX's CI colors. 
 #' 
-#' @param number a numeric \code{(1-9)} with the number of colors
+#' @param number a numeric with the number of colors.
+#'  The default is the length of \code{basecolors}.
 #' @param reorder a boolean if \code{TRUE} the order will be sampled,
 #'  which can make groups next to each other a little bit more separable.
+#' @param basecolors a numeric vector with the used color indices of 
+#' \code{\link{sci_palette()}}.
 #' 
 #' @importFrom grDevices rgb colorRampPalette
 #' 
@@ -16,11 +19,12 @@
 #' @examples
 #' number <- 16
 #' plot(data.frame(x =  1:number, y = 1), 
-#'      col = statworx_palette(number, TRUE),
+#'      col = statworx_palette(number, FALSE),
 #'      pch = 16, cex = 5)
 #'
-statworx_palette <- function(number = 4,
-                             reorder = FALSE) {
+statworx_palette <- function(number = length(basecolors),
+                             reorder = FALSE,
+                             basecolors = c(1,2,3,5,10)) {
   # check input format
   if (!is.numeric(number)) {
     stop("number needs to be numeric")
@@ -40,7 +44,7 @@ statworx_palette <- function(number = 4,
   }
   
   #number <- 20
-  getPalette <- grDevices::colorRampPalette(as.vector(sci_palette())[c(1,2,3,5)])
+  getPalette <- grDevices::colorRampPalette(as.vector(sci_palette())[basecolors])
   out <- getPalette(number)
   
   if (reorder) {