+++ NEWS.md
@@ -1,4 +1,4 @@
-## version 0.2.1
+## version 0.2.2
 
 ---
 
@@ -6,6 +6,17 @@
 ### Bugfixes
 
 - rename prop to mean in count_na
+
+
+## version 0.2.1
+
+---
+
+
+### update function
+
+- changed color names and added a red color to the palette
+- removed 'number' form sci_palette()
 
 
 ## version 0.2.0